lib update

+++ src/main/java/com/worldplugins/vip/command/vip/RemovePendingVips.java
@@ -1,36 +1,37 @@
 package com.worldplugins.vip.command.vip;
 
-import com.worldplugins.lib.command.CommandModule;
-import com.worldplugins.lib.command.annotation.ArgsChecker;
-import com.worldplugins.lib.command.annotation.Command;
-import com.worldplugins.lib.extension.GenericExtensions;
 import com.worldplugins.vip.database.pending.PendingVipRepository;
-import com.worldplugins.vip.extension.ResponseExtensions;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.experimental.ExtensionMethod;
+import me.post.lib.command.CommandModule;
+import me.post.lib.command.annotation.Command;
 import org.bukkit.command.CommandSender;
+import org.jetbrains.annotations.NotNull;
 
-@ExtensionMethod({
-    ResponseExtensions.class,
-    GenericExtensions.class
-})
+import static com.worldplugins.vip.Response.respond;
+import static me.post.lib.util.Pairs.to;
 
-@RequiredArgsConstructor
 public class RemovePendingVips implements CommandModule {
-    private final @NonNull PendingVipRepository pendingVipRepository;
+    private final @NotNull PendingVipRepository pendingVipRepository;
 
-    @Command(
-        name = "vip removerpendentes",
-        permission = "worldvip.removerpendentes",
-        argsChecks = {@ArgsChecker(size = 1)},
-        usage = "&cArgumentos invalidos. Digite /vip removerpendentes <jogador>"
-    )
+    public RemovePendingVips(@NotNull PendingVipRepository pendingVipRepository) {
+        this.pendingVipRepository = pendingVipRepository;
+    }
+
+    @Command(name = "vip removerpendentes")
     @Override
-    public void execute(@NonNull CommandSender sender, @NonNull String[] args) {
+    public void execute(@NotNull CommandSender sender, @NotNull String[] args) {
+        if (!sender.hasPermission("worldvip.removerpendentes")) {
+            respond(sender, "Remover-pendentes-permissoes");
+            return;
+        }
+
+        if (args.length != 1) {
+            respond(sender, "Remover-pendentes-uso");
+            return;
+        }
+
         pendingVipRepository.removePendings(args[0]);
-        sender.respond("Vip-pendentes-removidos", message -> message.replace(
-            "@jogador".to(args[0])
+        respond(sender, "Vip-pendentes-removidos", message -> message.replace(
+            to("@jogador", args[0])
         ));
     }
 }
